Move `QuestionFilters` component
